Ui update
加上一些簡單的UI部件
