+++ Assets/Scripts/Manager/BattleManagers/BattleMainManager.cs
@@ -4,15 +4,20 @@ using UnityEngine;
 
 public class BattleMainManager : MonoSingleton<BattleMainManager>
 {
+    [Header("開場抽幾張卡")]
+    public int initialDrawCard;
 
     void Start()
     {
-        
+        BattleStart();
     }
 
     public void BattleStart()
     {
+        
         TurnPhaseManager.instance.StartGame();
+        BattleDeckManager.instance.ShuffleDeck();
+        BattleDeckManager.instance.DrawCard(initialDrawCard);
     }
 
 }
