+++ Assets/Scripts/Manager/BattleManagers/BattlePlayerDataManager.cs
@@ -1,0 +1,23 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class BattlePlayerDataManager : MonoSingleton<BattlePlayerDataManager>,IDataPersistence
+{
+    public PlayerData playerData;
+
+    public void LoadData(GameData data)
+    {
+        
+    }
+
+    public void SaveData(ref GameData data)
+    {
+        
+    }
+
+    public void ShowPlayerStatus()
+    {
+
+    }
+}
