+++ Assets/Scripts/Manager/BattleManagers/BattlePlayerDataManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73c70f434c77a42b8a362b69bada91c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
